third_party/blink/web_tests/platform/generic/fast/forms/input-readonly-select-expected.txt

double clicking on normal input
PASS selectEventFiredOnInput is true


double clicking on readonly input
PASS selectEventFiredOnReadonlyInput is true


double clicking on disabled input
PASS selectEventFiredOnDisabledInput is true


double clicking on normal textarea
PASS selectEventFiredOnTextarea is true


double clicking on readonly textarea
PASS selectEventFiredOnReadonlyTextarea is true


double clicking on disabled textarea
PASS selectEventFiredOnDisabledTextarea is true


PASS successfullyParsed is true

TEST COMPLETE
